lab5_grpc_n_javarmi/rpc_grpc_byogrpc/pyproject.toml
[project]
name = "rpc-grpc-byogrpc"
version = "0.1.0"
description = "gRPC - Build Your Own gRPC Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio==1.66.2",
    "grpcio-tools==1.66.2",
]
